update: splitting the simulator file into two.
